+++ CHANGES.md
@@ -11,6 +11,11 @@
     Layout modifier that, if only a single window is on screen, places that window
     in the middle of the screen.
 
+  * `XMonad.Util.ActionQueue`
+
+    Put XMonad actions in the queue to be executed every time the
+    `logHook` (or, alternatively, a hook of your choice) runs.
+
 ### Bug Fixes and Minor Changes
 
   * `XMonad.Prompt`